+++ src/main/java/ru/hh/resumebuilderbot/BotBody.java
@@ -1,6 +1,7 @@
 package ru.hh.resumebuilderbot;
 
 import ru.hh.resumebuilderbot.MessageHandler.MessageHandler;
+import ru.hh.resumebuilderbot.QuestionGenerator.QuestionGenerator;
 
 import java.util.Map;
