+++ src/main/java/ru/hh/resumebuilderbot/MessageHandler/MessageHandler.java
@@ -1,7 +1,7 @@
 package ru.hh.resumebuilderbot.MessageHandler;
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.QuestionGenerator;
+import ru.hh.resumebuilderbot.QuestionGenerator.QuestionGenerator;
 
 public interface MessageHandler {
     public QuestionGenerator handle(Answer answer);
