+++ src/main/java/ru/hh/resumebuilderbot/MessageHandler/UnknownMessageHandler.java
@@ -2,8 +2,8 @@ package ru.hh.resumebuilderbot.MessageHandler;
 
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.FixedQuestionGenerator;
-import ru.hh.resumebuilderbot.QuestionGenerator;
+import ru.hh.resumebuilderbot.QuestionGenerator.FixedQuestionGenerator;
+import ru.hh.resumebuilderbot.QuestionGenerator.QuestionGenerator;
 import ru.hh.resumebuilderbot.TextsStorage;
 
 public class UnknownMessageHandler implements MessageHandler {
