--- src/main/java/ru/hh/resumebuilderbot/QuestionGenerator.java
@@ -1,5 +1,0 @@
-package ru.hh.resumebuilderbot;
-
-public interface QuestionGenerator {
-    public Question generate(ChatId chatId);
-}
