+++ src/main/java/ru/hh/resumebuilderbot/QuestionGenerator/QuestionGenerator.java
@@ -1,0 +1,8 @@
+package ru.hh.resumebuilderbot.QuestionGenerator;
+
+import ru.hh.resumebuilderbot.ChatId;
+import ru.hh.resumebuilderbot.Question;
+
+public interface QuestionGenerator {
+    public Question generate(ChatId chatId);
+}
