+++ src/main/java/ru/hh/resumebuilderbot/QuestionGenerator/ShowAllQuestionGenerator.java
@@ -1,4 +1,7 @@
-package ru.hh.resumebuilderbot;
+package ru.hh.resumebuilderbot.QuestionGenerator;
+
+import ru.hh.resumebuilderbot.ChatId;
+import ru.hh.resumebuilderbot.Question;
 
 public class ShowAllQuestionGenerator implements QuestionGenerator {
     @Override
